requirements.txt
cmsplugin-filer==0.10.2
Django==1.8.5
django-appconf==1.0.1
django-classy-tags==0.6.2
django-cms==3.1.3
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-filer==0.9.12
django-mptt==0.7.4
django-polymorphic==0.7.2
django-reversion==1.9.3
django-sekizai==0.8.2
Django-Select2==4.3.2
django-taggit==0.17.3
django-treebeard==3.0
djangocms-admin-style==0.2.8
djangocms-googlemap==0.3
djangocms-link==1.7.1
djangocms-text-ckeditor==2.6.0
easy-thumbnails==2.2
html5lib==0.9999999
Pillow==3.0.0
psycopg2==2.6.1
python-logstash==0.4.5
python-memcached==1.57
pytz==2015.7
six==1.10.0
sqlparse==0.1.18
Unidecode==0.4.18
